Created setup.py, in preparation to upload to PyPi.
Modified the structure to use our package (`filecrypt`) instead of
polluting the "root" namespace.

Created a binary runner (`run`) to launch the script, for testing
development (this is not installed in users installs - the
`console_script`, `encrypt` should be used instead).

Created the `setup.py` and enabled the conversion of the README.md
to RST (which is what PyPi understands) using `pandoc`.

Tested locally, next step is to test on PyPi and then create a release.